+++ core-bundle/src/Resources/contao/elements/ContentDownloads.php
@@ -10,6 +10,8 @@
 
 namespace Contao;
 
+use Contao\CoreBundle\Exception\PageNotFoundException;
+
 /**
  * Front end content element "downloads".
  *
@@ -68,8 +70,8 @@ class ContentDownloads extends ContentElement
 
 		$file = \Input::get('file', true);
 
-		// Send the file to the browser and do not send a 404 header (see #4632)
-		if ($file != '' && !preg_match('/^meta(_[a-z]{2})?\.txt$/', basename($file)))
+		// Send the file to the browser (see #4632 and #8375)
+		if ($file != '' && (!isset($_GET['cid']) || \Input::get('cid') == $this->id))
 		{
 			while ($this->objFiles->next())
 			{
@@ -77,6 +79,11 @@ class ContentDownloads extends ContentElement
 				{
 					\Controller::sendFileToBrowser($file);
 				}
+			}
+
+			if (isset($_GET['cid']))
+			{
+				throw new PageNotFoundException('Invalid file name');
 			}
 
 			$this->objFiles->reset();
@@ -141,12 +148,17 @@ class ContentDownloads extends ContentElement
 				$strHref = \Environment::get('request');
 
 				// Remove an existing file parameter (see #5683)
-				if (preg_match('/(&(amp;)?|\?)file=/', $strHref))
+				if (isset($_GET['file']))
 				{
 					$strHref = preg_replace('/(&(amp;)?|\?)file=[^&]+/', '', $strHref);
 				}
 
-				$strHref .= (strpos($strHref, '?') !== false ? '&amp;' : '?') . 'file=' . \System::urlEncode($objFiles->path);
+				if (isset($_GET['cid']))
+				{
+					$strHref = preg_replace('/(&(amp;)?|\?)cid=\d+/', '', $strHref);
+				}
+
+				$strHref .= (strpos($strHref, '?') !== false ? '&amp;' : '?') . 'file=' . \System::urlEncode($objFiles->path) . '&amp;cid=' . $this->id;
 
 				// Add the image
 				$files[$objFiles->path] = array